use a main file to start a working server; unit tests only for the methods not the setup
